% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternHelperFunctions.R
\name{convertSizeToCartesian}
\alias{convertSizeToCartesian}
\title{Converts a point size to Cartesian units}
\usage{
convertSizeToCartesian(size, scale, axis)
}
\arguments{
\item{size}{Size of the pointGrob}

\item{scale}{Minimum and maximum values of given scale}

\item{axis}{Cartesian axis: 'x' or 'y'}
}
\value{
Radius of the point in Cartesian units
}
\description{
Given a point size and the scale of an axis, this function is able to
convert the radius of the pointGrob being created to its width in
Cartesian units.  This function is used to ensure the patterns extend
a whole point or region no matter the point size.
}
