% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternHelperFunctions.R
\name{countGridPoints}
\alias{countGridPoints}
\title{Creates a 2D frequency matrix}
\usage{
countGridPoints(x, y, xDiff, yDiff, n = 25)
}
\arguments{
\item{x}{x-coordinates of the group}

\item{y}{y-coordinates of the group}

\item{xDiff}{Range of the x-axis in the scatterplot}

\item{yDiff}{Range of the y-axis in the scatterplot}

\item{n}{Size of the square matrix}
}
\value{
List: x & y bins of the grid, 2D frequency matrix, and dataframe matching each point to its corresponding grid in the matrix
}
\description{
This function creates a matrix of size n by n.  Each grid of the
matrix corresponds to a range of x coordinates and y coordinates
in the scatterplot.  The amount of points from a specific group
in each grid are counted.  Increasing the grid size, or n, increases
the precision of the patterns.
}
