% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHorizontal.R
\name{drawHorizontal}
\alias{drawHorizontal}
\title{Draws the horizontal line pattern}
\usage{
drawHorizontal(
  gridOutput,
  density = NULL,
  pointSize,
  xDiff,
  yDiff,
  rotatedxDiff,
  rotatedyDiff,
  sparsePoints = NULL
)
}
\arguments{
\item{gridOutput}{List: x & y bins of the grid, 2D frequency matrix, and dataframe matching each point to its corresponding grid in the matrix}

\item{density}{Density of the pattern to be drawn.  Must be 1 or less.  Skips drawing the pattern every n rows.}

\item{pointSize}{Point size of the pattern.}

\item{xDiff}{x-coordinate range of the plot.}

\item{yDiff}{y-coordinate range of the plot.}

\item{rotatedxDiff}{Rotated x-coordinate range of the plot.}

\item{rotatedyDiff}{Rotated y-coordinate range of the plot.}

\item{sparsePoints}{Logical Vector denoting points annotated as sparse.  If NULL, default sparsity detector will be used to annotate sparse points in dataset.}
}
\value{
Dataframe with 4 columns defining the starting and ending coordinates of each line to be drawn.
}
\description{
Using the 2D frequency matrix from countGridPoints(), this function goes row by row
checking if each grid contains a point.  The function records the leftmost and rightmost points in a
consecutive sequence of grids with points.  The function also records points in a sparse grid: defined
as being a grid where grids n-rows or n-columns away do not contain any points.
}
