% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positiveDiagnol.R
\name{drawPositiveDiagnol}
\alias{drawPositiveDiagnol}
\title{Draws the positive diagnol pattern}
\usage{
drawPositiveDiagnol(
  gridOutput,
  density = 1/2,
  sparsity = 5,
  size = 1,
  xDiff,
  yDiff
)
}
\arguments{
\item{gridOutput}{List: x & y bins of the grid, 2D frequency matrix, and dataframe matching each point to its corresponding grid in the matrix}

\item{density}{Density of the pattern to be drawn.  Must be 1 or less.  Skips drawing the pattern every n diagnols.}

\item{sparsity}{Threshold for defining a sparse grid.}

\item{size}{Point size of the pattern.}

\item{xDiff}{x-coordinate range of the plot.}

\item{yDiff}{y-coordinate range of the plot.}
}
\value{
Dataframe with 4 columns defining the starting and ending coordinates of each line to be drawn.
}
\description{
Using the 2D frequency matrix from countGridPoints(), this function goes diagnol by diagnol
checking if each grid contains a point.  The function records the lowest and highest points in a
consecutive sequence of grids with points.  The function also records points in a sparse grid: defined
as being a grid where grids n-rows or n-columns away do not contain any points.
}
