% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendHelperFunctions.R
\docType{data}
\name{imagePoints}
\alias{imagePoints}
\title{Creates custom ggplot2 object}
\format{
An object of class \code{imagePoints} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
imagePoints
}
\value{
List of grob objects
}
\description{
ggplot2 object that transfers the pattern aesthetics (e.g. color, lineColor, etc.) to render the legend icon.
}
\keyword{datasets}
