% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendHelperFunctions.R
\name{legendIcon}
\alias{legendIcon}
\title{Creating a Legend Icon}
\usage{
legendIcon(
  color,
  lineColor = "black",
  lineType = "solid",
  patternType,
  lineWidth = 1.5,
  angle = NULL
)
}
\arguments{
\item{color}{Background color of the icon}

\item{lineColor}{Color of the pattern lines}

\item{lineType}{Line type of pattern lines (e.g. dashed, dotted, etc.)}

\item{patternType}{Pattern Type of icon (e.g. vertical, horizontal, etc.)}

\item{lineWidth}{Width of the line}

\item{angle}{Vector of the line angles for a given pattern}
}
\value{
List of grob objects
}
\description{
This function creates the legend icons for the patterns inputted to scatterHatch.
The legend icons are rendered using grob raster objects in the grid package.
}
