% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdacData.R
\docType{data}
\name{pdacData}
\alias{pdacData}
\title{Tissue-CyCIF data of Pancreatic Ductal Adenocarcinoma (PDAC)}
\format{
A data frame with 158159 rows and 41 variables:
}
\source{
\url{http://spatialgiotto.rc.fas.harvard.edu/giotto.cycif.html}
}
\usage{
pdacData
}
\description{
A dataset containing antibody expression of a human tissue sample with
PDAC at a single-cell resolution used in Lin et al.
}
\keyword{datasets}
