% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateCoords.R
\name{rotateCoords}
\alias{rotateCoords}
\title{Finds the equivalent cartesian coordinates for points in a rotated coordinate system}
\usage{
rotateCoords(x, y, angle)
}
\arguments{
\item{x}{Vector of the x coordinates}

\item{y}{Vector of the y coordinates}

\item{angle}{Angle to rotate the coordinate system by (in degrees)}
}
\value{
Dataframe with 2 columns specifying the transformed coordinates for each point
}
\description{
Finds the equivalent cartesian coordinates for points in a rotated coordinate system
}
