% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterHatch.R
\name{scatterHatch}
\alias{scatterHatch}
\title{Creates a scatterplot with hatched patterns.}
\usage{
scatterHatch(
  data,
  x,
  y,
  factor,
  factorName,
  pointSize = 1.5,
  gridSize = NULL,
  patternList = list(list(pattern = "horizontal"), list(pattern = "vertical"),
    list(pattern = "positiveDiagonal"), list(pattern = "negativeDiagonal")),
  colorPalette = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00",
    "#CC79A7")
)
}
\arguments{
\item{data}{Dataset (what normally would be passed to ggplot2)}

\item{x}{x-coordinates of the plot}

\item{y}{y-coordinates of the plot}

\item{factor}{Factor vector that defines what group each point is part of}

\item{factorName}{Title of the legend}

\item{pointSize}{Point size for the scatterplot}

\item{gridSize}{Controls the precision of the hatched pattern.  Larger values correspond to greater precision.  Default follows a exponential decay function based on point size.}

\item{patternList}{Aesthetics to be passed for each pattern (must be a list where each element has a named pattern)}

\item{colorPalette}{Colors to be used for each group.  Default is color-blind friendly.}
}
\value{
ggplot2 object of scatterplot with hatched patterns for each group.
}
\description{
This function creates a scatterplot with hatched patterns by using geom_segment().
Also creates the legend to represent each hatched group.  The aesthetics defining each pattern,
like the lineType and lineColor are all passed to geom_segment().
}
