% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsityAnnotate.R
\name{sparsityAnnotate}
\alias{sparsityAnnotate}
\title{Determines sparse points in a group}
\usage{
sparsityAnnotate(pointsToGrid, pointSize, xDiff, yDiff, scale)
}
\arguments{
\item{pointsToGrid}{Dataframe specifying every point in a group and the grid it belongs to}

\item{pointSize}{Point size of the plot.}

\item{xDiff}{x-coordinate range of the plot.}

\item{yDiff}{y-coordinate range of the plot.}

\item{scale}{If the sparsity of the pattern calculated is dependent which axes}
}
\value{
Dataframe of each point in a group, which grid it belongs to, distance from fifth nearest point, and whether a point is sparse or not.
}
\description{
Determines which points within a given group are sparse
and which are in small clusters.  Outlying points are determined if
fifth closest point is a point away.  Small clusters are determined
if twentieth closest point is 2 points away.
}
