% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendHelperFunctions.R
\name{geom_imagePoint}
\alias{geom_imagePoint}
\title{Creates custom ggplot2 layer with geom mentioned above}
\usage{
geom_imagePoint(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings}

\item{data}{Data to be displayed in this layer}

\item{stat}{Statistical transformation to use on the data for this layer}

\item{position}{Position adjustment}

\item{na.rm}{if missing values are to be removed}

\item{show.legend}{if layer be included in the legends}

\item{inherit.aes}{if layer should inherit behavior}

\item{...}{all remaining params passed to create a layer}
}
\value{
List of grob objects
}
\description{
Creates custom ggplot2 layer with geom mentioned above
}
