% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendHelperFunctions.R
\name{geom_imagePoint}
\alias{geom_imagePoint}
\title{Creates custom ggplot2 layer with geom mentioned above}
\usage{
geom_imagePoint(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\value{
List of grob objects
}
\description{
ggplot2 object that transfers the pattern aesthetics (e.g. color, lineColor, etc.) to render the legend icon.
}
