% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendHelperFunctions.R
\name{legendIcon}
\alias{legendIcon}
\title{Creating a Legend Icon}
\usage{
legendIcon(
  color,
  lineColor = "black",
  lineType = "solid",
  patternType,
  lineWidth = 1.5,
  lineAlpha = 0.4,
  angle
)
}
\arguments{
\item{color}{Background color of the icon}

\item{lineColor}{Color of the pattern lines}

\item{lineType}{Line type of pattern lines (e.g. dashed, dotted, etc.)}

\item{patternType}{Pattern Type of icon (e.g. vertical, horizontal, etc.)}

\item{lineWidth}{Width of the line}

\item{angle}{Vector of the line angles for a given pattern}

\item{alpha}{Transparency of the line}
}
\value{
List of grob objects
}
\description{
This function creates the legend icons for the four distinct patterns supported at the moment:
horizontal, vertical, positiveDiagnol, and negativeDiagnol.  The legend icons are rendered using
grob raster objects in the grid package.  Before the grid method, the icons were drawn using
image_draw in the Magick package.  The Magick method was proven impractical because it created
an issue with the graphic device preventing ggsave from rendering the final plot.
}
