% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdacData.R
\docType{data}
\name{pdacData}
\alias{pdacData}
\title{Pancreatic Dual Adenocarcionma (PDAC) tissue-CyCIF dataset}
\format{
An object of class \code{data.frame} with 158159 rows and 41 columns.
}
\source{
\href{http://spatial.rc.fas.harvard.edu/spatialgiotto/giotto.dataset.html}{Giotto Datasets}
}
\usage{
data(pdacData)
}
\description{
Single-cell resolution of 21 proteins and 3 cellular compartment in a human
pancreatic dual adenocarcionma sample that spans across three distinct tissues:
the pancreas, small intestine, and tumor.  160,000 cells profiled.
}
\references{
Lin et al. (2018) eLife 2018;7:e31657
(\href{https://pubmed.ncbi.nlm.nih.gov/29993362/}{PubMed})
}
\keyword{datasets}
